# app/src/main/webui/api/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "contact" : {
      "name" : "ldclrcq"
    },
    "title" : "Homelab Manager API",
    "version" : "0.0.1"
  },
...
  "components" : {
    "schemas" : {
      "AdminProblemWorkflowDto" : {
        "type" : "object",
        "required" : [ "username", "workflow" ],
        "properties" : {
          "username" : {
            "type" : "string"
          },
          "workflow" : {
            "$ref" : "#/components/schemas/ProblemWorkflowDto"
          }
        }
      },
...
      "CreateWorkflowRequest" : {
        "type" : "object",
        "required" : [ "mediaType" ],
        "properties" : {
          "mediaType" : {
            "type" : [ "string", "null" ]
          }
        }
      },
...
      "GrabReleaseRequest" : {
        "type" : "object",
        "required" : [ "guid", "indexerId", "title", "indexer", "quality", "size" ],
        "properties" : {
          "guid" : {
            "type" : [ "string", "null" ]
          },
          "indexerId" : {
            "type" : [ "integer", "null" ],
            "format" : "int32"
          },
          "title" : {
            "type" : [ "string", "null" ]
          },
          "indexer" : {
            "type" : [ "string", "null" ]
          },
          "quality" : {
            "type" : [ "string", "null" ]
          },
          "size" : {
            "type" : [ "integer", "null" ],
            "format" : "int64"
          }
        }
      },
      "GrabbedRelease" : {
        "type" : "object",
        "properties" : {
          "guid" : {
            "type" : [ "string", "null" ]
          },
          "indexerId" : {
            "type" : [ "integer", "null" ],
            "format" : "int32"
          },
          "indexer" : {
            "type" : [ "string", "null" ]
          },
          "title" : {
            "type" : [ "string", "null" ]
          },
          "quality" : {
            "type" : [ "string", "null" ]
          },
          "size" : {
            "type" : [ "integer", "null" ],
            "format" : "int64"
          }
        }
      },
...
      "LocalDateTime" : {
        "type" : "string",
        "format" : "date-time",
        "examples" : [ "2022-03-10T12:15:50" ]
      },
...
      "MediaSnapshot" : {
        "type" : "object",
        "properties" : {
          "title" : {
            "type" : [ "string", "null" ]
          },
          "year" : {
            "type" : [ "integer", "null" ],
            "format" : "int32"
          },
          "posterUrl" : {
            "type" : [ "string", "null" ]
          },
          "overview" : {
            "type" : [ "string", "null" ]
          },
          "currentQuality" : {
            "type" : [ "string", "null" ]
          },
          "currentLanguages" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
...
      "ProblemReleaseDto" : {
        "type" : "object",
        "required" : [ "guid", "indexerId", "indexer", "title", "quality", "size", "age", "seeders", "leechers", "protocol", "rejected", "rejections", "isFrench", "isRecommended" ],
        "properties" : {
          "guid" : {
            "type" : "string"
          },
          "indexerId" : {
            "type" : "integer",
            "format" : "int32"
          },
          "indexer" : {
            "type" : [ "string", "null" ]
          },
          "title" : {
            "type" : "string"
          },
          "quality" : {
            "type" : [ "string", "null" ]
          },
          "size" : {
            "type" : [ "integer", "null" ],
            "format" : "int64"
          },
          "age" : {
            "type" : [ "integer", "null" ],
            "format" : "int32"
          },
          "seeders" : {
            "type" : [ "integer", "null" ],
            "format" : "int32"
          },
          "leechers" : {
            "type" : [ "integer", "null" ],
            "format" : "int32"
          },
          "protocol" : {
            "type" : [ "string", "null" ]
          },
          "rejected" : {
            "type" : "boolean"
          },
          "rejections" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "isFrench" : {
            "type" : "boolean"
          },
          "isRecommended" : {
            "type" : "boolean"
          }
        }
      },
...
      "ProblemWorkflowDto" : {
        "type" : "object",
        "required" : [ "id", "mediaType", "status", "currentStep", "problemType", "description", "media", "grabbedRelease", "createdAt", "updatedAt", "completedAt" ],
        "properties" : {
          "id" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "mediaType" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "currentStep" : {
            "type" : "string"
          },
          "problemType" : {
            "type" : [ "string", "null" ]
          },
          "description" : {
            "type" : [ "string", "null" ]
          },
          "media" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/MediaSnapshot"
            }, {
              "type" : "null"
            } ]
          },
          "grabbedRelease" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/GrabbedRelease"
            }, {
              "type" : "null"
            } ]
          },
          "createdAt" : {
            "$ref" : "#/components/schemas/LocalDateTime"
          },
          "updatedAt" : {
            "$ref" : "#/components/schemas/LocalDateTime"
          },
          "completedAt" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/LocalDateTime"
            }, {
              "type" : "null"
            } ]
          }
        }
      },
...
      "SelectMovieRequest" : {
        "type" : "object",
        "required" : [ "radarrMovieId" ],
        "properties" : {
          "radarrMovieId" : {
            "type" : [ "integer", "null" ],
            "format" : "int32"
          }
        }
      },
      "SelectProblemRequest" : {
        "type" : "object",
        "required" : [ "problemType", "description" ],
        "properties" : {
          "problemType" : {
            "type" : [ "string", "null" ]
          },
          "description" : {
            "type" : [ "string", "null" ]
          }
        }
      },
      "SelectSeriesRequest" : {
        "type" : "object",
        "required" : [ "sonarrSeriesId" ],
        "properties" : {
          "sonarrSeriesId" : {
            "type" : [ "integer", "null" ],
            "format" : "int32"
          }
        }
      },
...
      "UUID" : {
        "type" : "string",
        "format" : "uuid",
        "pattern" : "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
      },
...
    }
  },
  "paths" : {
...
    "/api/admin/problems/workflows" : {
      "get" : {
        "tags" : [ "Problems" ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/AdminProblemWorkflowDto"
                  }
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "summary" : "List Workflows",
        "security" : [ {
          "SecurityScheme" : [ "admin" ]
        } ]
      }
    },
    "/api/admin/problems/workflows/{id}" : {
      "get" : {
        "tags" : [ "Problems" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminProblemWorkflowDto"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "summary" : "Get Workflow",
        "security" : [ {
          "SecurityScheme" : [ "admin" ]
        } ]
      },
      "delete" : {
        "tags" : [ "Problems" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Workflow supprimé"
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "summary" : "Delete Workflow",
        "security" : [ {
          "SecurityScheme" : [ "admin" ]
        } ]
      }
    },
    "/api/admin/problems/workflows/{id}/abandon" : {
      "post" : {
        "tags" : [ "Problems" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemWorkflowDto"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "summary" : "Abandon Workflow",
        "security" : [ {
          "SecurityScheme" : [ "admin" ]
        } ]
      }
    },
    "/api/admin/problems/workflows/{id}/grab" : {
      "post" : {
        "tags" : [ "Problems" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/GrabReleaseRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemWorkflowDto"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Grab Release",
        "security" : [ {
          "SecurityScheme" : [ "admin" ]
        } ]
      }
    },
    "/api/admin/problems/workflows/{id}/movie" : {
      "post" : {
        "tags" : [ "Problems" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SelectMovieRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemWorkflowDto"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Select Movie",
        "security" : [ {
          "SecurityScheme" : [ "admin" ]
        } ]
      }
    },
    "/api/admin/problems/workflows/{id}/problem" : {
      "post" : {
        "tags" : [ "Problems" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SelectProblemRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemWorkflowDto"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Select Problem",
        "security" : [ {
          "SecurityScheme" : [ "admin" ]
        } ]
      }
    },
    "/api/admin/problems/workflows/{id}/releases" : {
      "get" : {
        "tags" : [ "Problems" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ProblemReleaseDto"
                  }
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "summary" : "List Releases",
        "security" : [ {
          "SecurityScheme" : [ "admin" ]
        } ]
      }
    },
    "/api/admin/problems/workflows/{id}/resolve" : {
      "post" : {
        "tags" : [ "Problems" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemWorkflowDto"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "summary" : "Resolve Workflow",
        "security" : [ {
          "SecurityScheme" : [ "admin" ]
        } ]
      }
    },
    "/api/admin/problems/workflows/{id}/series" : {
      "post" : {
        "tags" : [ "Problems" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SelectSeriesRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemWorkflowDto"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Select Series",
        "security" : [ {
          "SecurityScheme" : [ "admin" ]
        } ]
      }
    },
...
    "/api/problems/workflows" : {
      "get" : {
        "tags" : [ "Problems" ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ProblemWorkflowDto"
                  }
                }
              }
            }
          }
        },
        "summary" : "List Workflows"
      },
      "post" : {
        "tags" : [ "Problems" ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateWorkflowRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemWorkflowDto"
                }
              }
            },
            "description" : "Created"
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Create Workflow"
      }
    },
    "/api/problems/workflows/{id}" : {
      "get" : {
        "tags" : [ "Problems" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemWorkflowDto"
                }
              }
            }
          }
        },
        "summary" : "Get Workflow"
      }
    },
    "/api/problems/workflows/{id}/abandon" : {
      "post" : {
        "tags" : [ "Problems" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemWorkflowDto"
                }
              }
            }
          }
        },
        "summary" : "Abandon Workflow"
      }
    },
    "/api/problems/workflows/{id}/grab" : {
      "post" : {
        "tags" : [ "Problems" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/GrabReleaseRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemWorkflowDto"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Grab Release"
      }
    },
    "/api/problems/workflows/{id}/movie" : {
      "post" : {
        "tags" : [ "Problems" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SelectMovieRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemWorkflowDto"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Select Movie"
      }
    },
    "/api/problems/workflows/{id}/problem" : {
      "post" : {
        "tags" : [ "Problems" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SelectProblemRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemWorkflowDto"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Select Problem"
      }
    },
    "/api/problems/workflows/{id}/releases" : {
      "get" : {
        "tags" : [ "Problems" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ProblemReleaseDto"
                  }
                }
              }
            }
          }
        },
        "summary" : "List Releases"
      }
    },
    "/api/problems/workflows/{id}/resolve" : {
      "post" : {
        "tags" : [ "Problems" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemWorkflowDto"
                }
              }
            }
          }
        },
        "summary" : "Resolve Workflow"
      }
    },
    "/api/problems/workflows/{id}/series" : {
      "post" : {
        "tags" : [ "Problems" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SelectSeriesRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemWorkflowDto"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Select Series"
      }
    },
...
  },
  "servers" : [ {
    "url" : "http://localhost:8088",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8088",
    "description" : "Auto generated value"
  } ]
}