# app/src/main/webui/src/api/model/adminProblemWorkflowDto.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * Homelab Manager API
 * OpenAPI spec version: 0.0.1
 */
import type { ProblemWorkflowDto } from "./problemWorkflowDto";

export interface AdminProblemWorkflowDto {
  username: string;
  workflow: ProblemWorkflowDto;
}