# app/src/main/webui/src/api/model/mediaSnapshot.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * Homelab Manager API
 * OpenAPI spec version: 0.0.1
 */
import type { MediaSnapshotTitle } from "./mediaSnapshotTitle";
import type { MediaSnapshotYear } from "./mediaSnapshotYear";
import type { MediaSnapshotPosterUrl } from "./mediaSnapshotPosterUrl";
import type { MediaSnapshotOverview } from "./mediaSnapshotOverview";
import type { MediaSnapshotCurrentQuality } from "./mediaSnapshotCurrentQuality";

export interface MediaSnapshot {
  title?: MediaSnapshotTitle;
  year?: MediaSnapshotYear;
  posterUrl?: MediaSnapshotPosterUrl;
  overview?: MediaSnapshotOverview;
  currentQuality?: MediaSnapshotCurrentQuality;
  currentLanguages?: string[];
}