# app/src/main/webui/src/api/model/problemReleaseDto.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * Homelab Manager API
 * OpenAPI spec version: 0.0.1
 */
import type { ProblemReleaseDtoIndexer } from "./problemReleaseDtoIndexer";
import type { ProblemReleaseDtoQuality } from "./problemReleaseDtoQuality";
import type { ProblemReleaseDtoSize } from "./problemReleaseDtoSize";
import type { ProblemReleaseDtoAge } from "./problemReleaseDtoAge";
import type { ProblemReleaseDtoSeeders } from "./problemReleaseDtoSeeders";
import type { ProblemReleaseDtoLeechers } from "./problemReleaseDtoLeechers";
import type { ProblemReleaseDtoProtocol } from "./problemReleaseDtoProtocol";

export interface ProblemReleaseDto {
  guid: string;
  indexerId: number;
  indexer: ProblemReleaseDtoIndexer;
  title: string;
  quality: ProblemReleaseDtoQuality;
  size: ProblemReleaseDtoSize;
  age: ProblemReleaseDtoAge;
  seeders: ProblemReleaseDtoSeeders;
  leechers: ProblemReleaseDtoLeechers;
  protocol: ProblemReleaseDtoProtocol;
  rejected: boolean;
  rejections: string[];
  isFrench: boolean;
  isRecommended: boolean;
}