# app/src/main/webui/src/api/model/selectProblemRequest.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * Homelab Manager API
 * OpenAPI spec version: 0.0.1
 */
import type { SelectProblemRequestProblemType } from "./selectProblemRequestProblemType";
import type { SelectProblemRequestDescription } from "./selectProblemRequestDescription";

export interface SelectProblemRequest {
  problemType: SelectProblemRequestProblemType;
  description: SelectProblemRequestDescription;
}