# app/src/main/webui/src/api/model/selectSeriesRequest.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * Homelab Manager API
 * OpenAPI spec version: 0.0.1
 */
import type { SelectSeriesRequestSonarrSeriesId } from "./selectSeriesRequestSonarrSeriesId";

export interface SelectSeriesRequest {
  sonarrSeriesId: SelectSeriesRequestSonarrSeriesId;
}